moved test helpers to new file
